feat(core): add fragmentation filter
This is a test filter similar to sharded test execution that AOSP provides https://source.android.com/devices/tech/test_infra/tradefed/architecture/advanced/sharding

It is intended to be used in situations where it is not possible to connect multiple execution devices to marathon, e.g. CI setup that can schedule parallel jobs each containing a single execution device.

This is a dynamic programming technique, hence the results will be sub-optimal compared to connecting multiple devices to the same test run